requirements.txt
streamlit
streamlit-aggrid
numpy
matplotlib
pandas
streamlit-elements
streamlit-option-menu
streamlit-pandas-profiling
plotly
pymongo
tk
langchain
langchain-openai
langchain-community
langchain-experimental
langchain-core
datacompy
python-dotenv
